requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
llama-index>=0.11.0
llama-index-readers-file>=0.2.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-vector-stores-faiss>=0.2.0
sentence-transformers>=2.6.0
faiss-cpu>=1.7.4
pypdf>=4.0.0
transformers>=4.41.0
accelerate>=0.30.0
sentencepiece>=0.2.0
streamlit>=1.33.0
